web/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "soundlab-web"
version = "0.1.0"
description = "Sound Lab 統合Webインターフェース"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
]

[project.scripts]
soundlab-web = "soundlab_web.app:main"
